services/auth_service/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt
mysqlclient==2.1.1
django-cors-headers==4.0.0
social-auth-app-django==5.4.1
pika==1.3.1
requests==2.28.2
gunicorn==20.1.0
Pillow>=9.0.0
